requirements.txt
django==4.2.21
djangorestframework==3.16.0
tensorflow==2.19.0
numpy>=1.26.0,<2.0.0
pandas>=2.2.0
scikit-learn==1.3.2
requests==2.32.3 